1. use deepep load_acquire 2. rename ll to intranode
