+++ csrc/python/bind.cpp
@@ -35,8 +35,8 @@
 #endif
 
 #ifdef BUILD_OPS
-#include "ops/intra/all_gather_ll/all_gather_ll.h"
-#include "ops/intra/all_gather_ll/all_gather_ll_buffer.h"
+#include "ops/intra/all_gather_intra/all_gather_intra.h"
+#include "ops/intra/all_gather_intra/all_gather_intra_buffer.h"
 #endif
 
 using json = nlohmann::json;
